+++ Auxilium/Auxilium/API/Controllers/IndexController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Auxilium\Auxilium\API\Controllers;
+namespace Auxilium\API\Controllers;
 
-use Auxilium\Auxilium\API\Models\IndexModel;
-use Auxilium\Auxilium\API\Superclasses\APIController;
-use Auxilium\Auxilium\API\Superclasses\APIModel;
+use Auxilium\API\Models\IndexModel;
+use Auxilium\API\Superclasses\APIController;
+use Auxilium\API\Superclasses\APIModel;
 use Auxilium\DatabaseInteractions\GraphDatabaseConnection;
 use Auxilium\SessionHandling\Session;
 use Darksparrow\DeegraphInteractions\DataStructures\DataURL;
@@ -59,7 +59,7 @@ class IndexController extends APIController
 
 
         $regenerate_index = false;
-        $index_list = json_decode(file_get_contents(__DIR__ . "/../../../../indexes.json"), true, 512, JSON_THROW_ON_ERROR);
+        $index_list = json_decode(file_get_contents(__DIR__ . "/../../../indexes.json"), true, 512, JSON_THROW_ON_ERROR);
 
         if(!array_key_exists($index_id, $index_list))
         {
