+++ Auxilium/Auxilium/API/Controllers/JobStatisticsController.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Auxilium\Auxilium\API\Controllers;
+namespace Auxilium\API\Controllers;
 
-use Auxilium\Auxilium\API\Models\JobStatsModel;
-use Auxilium\Auxilium\API\Superclasses\APIController;
+use Auxilium\API\Models\JobStatsModel;
+use Auxilium\API\Superclasses\APIController;
 use JetBrains\PhpStorm\NoReturn;
 use OpenApi\Attributes\Get;
 use OpenApi\Attributes\JsonContent;
