+++ Auxilium/Auxilium/API/Controllers/LFSController.php
@@ -1,22 +1,16 @@
 <?php
 
-namespace Auxilium\Auxilium\API\Controllers;
+namespace Auxilium\API\Controllers;
 
-use Auxilium\Auxilium\API\Models\IndexModel;
-use Auxilium\Auxilium\API\Models\LFSModel;
-use Auxilium\Auxilium\API\Superclasses\APIController;
-use Auxilium\Auxilium\API\Superclasses\APIModel;
+use Auxilium\API\Models\LFSModel;
+use Auxilium\API\Superclasses\APIController;
+use Auxilium\API\Superclasses\APIModel;
 use Auxilium\Auxilium\AuxiliumLFSObject;
-use Auxilium\DatabaseInteractions\GraphDatabaseConnection;
-use Auxilium\SessionHandling\Session;
 use Auxilium\Utilities\Security;
-use Darksparrow\DeegraphInteractions\DataStructures\DataURL;
 use JetBrains\PhpStorm\NoReturn;
 use OpenApi\Attributes\Get;
-use OpenApi\Attributes\JsonContent;
 use OpenApi\Attributes\Post;
 use OpenApi\Attributes\Response;
-use RuntimeException;
 
 class LFSController extends APIController
 {
