+++ Auxilium/Auxilium/API/Models/DraftModel.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Auxilium\Auxilium\API\Models;
+namespace Auxilium\API\Models;
 
-use Auxilium\Auxilium\API\Superclasses\APIModel;
+use Auxilium\API\Superclasses\APIModel;
 use OpenApi\Attributes\Schema;
 
 #[Schema(
